+++ app/src/main/res/values/strings.xml
@@ -6,6 +6,7 @@
     <string name="menu_around">둘러보기</string>
     <string name="menu_shopping">쇼핑</string>
     <string name="menu_mypage">마이페이지</string>
+    <string name="btn_add_post">포스트 작성</string>
 
     <string name="back_pressed_toast">뒤로가기 버튼을 한 번 더 누르면 앱이 종료됩니다.</string>
 </resources>
