+++ core/common-ui/src/main/java/com/swmarastro/mykkumi/common_ui/post/PostImagesAdapter.kt
@@ -43,41 +43,67 @@ class PostImagesAdapter(
 
             binding.relativePinsOfImages.removeAllViews()
 
-            // pin이 이미지의 크기를 벗어나지 않도록 제한
             val parent = binding.imagePost
 
-            parent.viewTreeObserver.addOnGlobalLayoutListener (
-                object : ViewTreeObserver.OnGlobalLayoutListener {
-                    override fun onGlobalLayout() {
-                        if (parent.width != 0 && parent.height != 0 && parentWidth == 0) {
-                            Log.d("test", "width: ${parent.width}, height: ${parent.height}")
-                            Log.d("test", "${binding.relativePostImage.width}")
-                            if(parent.width > parent.height) {
-                                parentWidth = binding.relativePostImage.width
-                                parentHeight = parent.height * (binding.relativePostImage.width / parent.width)
-                            }
-                            else {
-                                parentWidth = parent.width * (binding.relativePostImage.width / parent.height)
-                                parentHeight = binding.relativePostImage.width
-                            }
+            // 이미지 세로 사이즈를 가로 사이즈와 동일하게 설정
+            parent.viewTreeObserver.addOnGlobalLayoutListener {
+                val width = parent.width
+                parent.layoutParams.height = width
+                parent.requestLayout()
 
-                            binding.imagePost.layoutParams.width = parentWidth
-                            binding.imagePost.layoutParams.height = parentHeight
-                            binding.imagePost.requestLayout()
+                parentWidth = parent.width
+                parentHeight = parent.height
 
-                            Log.d("test", "------ width: ${parent.width}, height: ${parent.height}")
-                            Log.d("test", "------22 width: ${parentWidth}, height: ${parentHeight}")
+//                if(parent.width > parent.height) {
+//                    parentWidth = binding.relativePostImage.width
+//                    parentHeight = parent.height * (binding.relativePostImage.width / parent.width)
+//                }
+//                else {
+//                    parentWidth = parent.width * (binding.relativePostImage.width / parent.height)
+//                    parentHeight = binding.relativePostImage.width
+//                }
 
-                            binding.relativePinsOfImages.layoutParams.width = parentWidth
-                            binding.relativePinsOfImages.layoutParams.height = parentHeight
-                            binding.relativePinsOfImages.requestLayout()
-                            notifyItemChanged(position)
-                        }
+                binding.relativePinsOfImages.layoutParams.width = parentWidth
+                binding.relativePinsOfImages.layoutParams.height = parentHeight
+                binding.relativePinsOfImages.requestLayout()
+                notifyItemChanged(position)
+            }
 
-                        parent.viewTreeObserver.removeOnGlobalLayoutListener(this)
-                    }
-                }
-            )
+//            // pin이 이미지의 크기를 벗어나지 않도록 제한
+//            val parent = binding.imagePost
+//
+//            parent.viewTreeObserver.addOnGlobalLayoutListener (
+//                object : ViewTreeObserver.OnGlobalLayoutListener {
+//                    override fun onGlobalLayout() {
+//                        if (parent.width != 0 && parent.height != 0 && parentWidth == 0) {
+//                            Log.d("test", "width: ${parent.width}, height: ${parent.height}")
+//                            Log.d("test", "${binding.relativePostImage.width}")
+//                            if(parent.width > parent.height) {
+//                                parentWidth = binding.relativePostImage.width
+//                                parentHeight = parent.height * (binding.relativePostImage.width / parent.width)
+//                            }
+//                            else {
+//                                parentWidth = parent.width * (binding.relativePostImage.width / parent.height)
+//                                parentHeight = binding.relativePostImage.width
+//                            }
+//
+//                            binding.imagePost.layoutParams.width = parentWidth
+//                            binding.imagePost.layoutParams.height = parentHeight
+//                            binding.imagePost.requestLayout()
+//
+//                            Log.d("test", "------ width: ${parent.width}, height: ${parent.height}")
+//                            Log.d("test", "------22 width: ${parentWidth}, height: ${parentHeight}")
+//
+//                            binding.relativePinsOfImages.layoutParams.width = parentWidth
+//                            binding.relativePinsOfImages.layoutParams.height = parentHeight
+//                            binding.relativePinsOfImages.requestLayout()
+//                            notifyItemChanged(position)
+//                        }
+//
+//                        parent.viewTreeObserver.removeOnGlobalLayoutListener(this)
+//                    }
+//                }
+//            )
 
             // 포스트에 핀 추가
             for(idx in 0..<item.pins.size) {
