+++ core/common-ui/src/main/res/layout/item_post_image_view_viewpager.xml
@@ -11,7 +11,7 @@
 
         <ImageView
             android:id="@+id/image_post"
-            android:layout_width="wrap_content"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:adjustViewBounds="true"
             android:layout_centerInParent="true"
