+++ feature/home/src/main/java/com/swmarastro/mykkumi/feature/home/HomeFragment.kt
@@ -1,10 +1,6 @@
 package com.swmarastro.mykkumi.feature.home
 
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
-import android.provider.Settings
-import android.util.Log
 import android.view.View
 import android.widget.ScrollView
 import android.widget.Toast
@@ -15,17 +11,14 @@ import androidx.navigation.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.viewpager2.widget.ViewPager2
 import com.swmarastro.mykkumi.common_ui.base.BaseFragment
-import com.swmarastro.mykkumi.common_ui.permission.ImagePermissionUtils
 import com.swmarastro.mykkumi.domain.entity.BannerListVO
 import com.swmarastro.mykkumi.domain.entity.HomePostItemVO
 import com.swmarastro.mykkumi.feature.home.banner.HomeBannerViewPagerAdapter
 import com.swmarastro.mykkumi.feature.home.databinding.FragmentHomeBinding
 import com.swmarastro.mykkumi.feature.home.post.PostListAdapter
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.launch
 import java.util.Timer
 import java.util.TimerTask
 
@@ -38,7 +31,6 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
 
     private lateinit var timer: Timer
 
-
     private var navController: NavController? = null
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -48,18 +40,6 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
 
         startAutoScroll()
         onClickBannerAll() // 배너 > + 버튼 선택 시 전체 리스트 페이지로 이동
-
-        // 포스트 작성 -> 로그인 유도 -> 이미지, 카메라 접근 권한 요청 -> 포스트 작성 페이지
-        binding.floatingBtnAddPost.setOnClickListener {
-            val intent = viewModel.navigateLogin()
-            if(intent == null) { // 로그인 됨
-                // 이미지 권한 요청하기
-                checkPermissionsAndProceed()
-            }
-            else { // 로그인 안 됨
-                startActivity(intent)
-            }
-        }
 
         // 로그아웃 테스트
         binding.btnShoppingCart.setOnClickListener {
@@ -187,20 +167,8 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
 //
 //        }
 
-        lifecycleScope.launch {
-            viewModel.postListUiState.collect {
-                Log.d("test", "test1")
-                postListAdapter.postList = it
-                Log.d("test", "test2")
-
-                if (viewModel.postCursor.value.isNullOrEmpty()) {
-                    binding.includeListLoading.visibility = View.GONE
-                }
-            }
-        }
-
-//        viewModel.postListUiState
-//            .onEach {
+//        lifecycleScope.launch {
+//            viewModel.postListUiState.collect {
 //                Log.d("test", "test1")
 //                postListAdapter.postList = it
 //                Log.d("test", "test2")
@@ -209,29 +177,17 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
 //                    binding.includeListLoading.visibility = View.GONE
 //                }
 //            }
-//            .launchIn(viewLifecycleOwner.lifecycleScope)
-    }
+//        }
 
-    private fun checkPermissionsAndProceed() {
-        if (ImagePermissionUtils.allPermissionsGranted(requireContext())) {
-            // 포스트 작성 페이지로 이동
-            viewModel.navigatePostEdit(navController)
-        } else if (ImagePermissionUtils.shouldShowRationale(requireActivity())) {
-            val intent = Intent(
-                Settings.ACTION_APPLICATION_DETAILS_SETTINGS,
-                Uri.parse("package:${requireContext().packageName}")
-            ).apply {
-                addCategory(Intent.CATEGORY_DEFAULT)
-                flags = Intent.FLAG_ACTIVITY_NEW_TASK
+        viewModel.postListUiState
+            .onEach {
+                postListAdapter.postList = it
+
+                if (viewModel.postCursor.value.isNullOrEmpty()) {
+                    binding.includeListLoading.visibility = View.GONE
+                }
             }
-
-            ImagePermissionUtils.showSettingsSnackbar(
-                requireActivity(), requireView(),
-                intent
-            )
-        } else {
-            ImagePermissionUtils.requestPermissions(requireActivity())
-        }
+            .launchIn(viewLifecycleOwner.lifecycleScope)
     }
 
     override fun onDestroyView() {
